uzbaigtas slegio daviklis MonitoringBlockV2 + antras interuptas susetupintas

+++ Code/Monitoringblock/MonitoringBlockV2/MonitoringBlockV2/Headers/PressureSensor.h
@@ -9,7 +9,7 @@
 #ifndef PRESSURESENSOR_H_
 #define PRESSURESENSOR_H_
 
-float GetPressure();
+int GetPressure();
